fixtures-service/requirements.txt
faker==2.0.3
more-itertools==7.2.0
python-dateutil==2.8.0
unidecode==1.1.1
